dynamic-filtering-form/requirements.txt
Django==4.0.4
django-filter==21.1
django-bootstrap4==22.1
